refactored action locations
