+++ frontend/src/store/actions/localStream.actions.js
@@ -1,23 +1,8 @@
-import {ADD_STREAM, CLEAR_ALL_STREAMS, REMOVE_STREAM, SET_LOCAL_USER_MEDIA_STREAM} from '../actionTypes';
+import {addStream} from './remoteStreams.actions';
 import {clearStreamTracks} from '../../utils';
+import {SET_LOCAL_USER_CAMERA_ENABLED, SET_LOCAL_USER_MEDIA_CONSTRAINTS, SET_LOCAL_USER_MEDIA_STREAM, SET_LOCAL_USER_MICROPHONE_ENABLED} from '../actionTypes';
 
-export const addStream = (socketId, stream) => (dispatch, getState) => { // TODO no need for a thunk, rename to setStream & SET_STREAM
-  if (stream) {
-    dispatch({ type: ADD_STREAM, payload: { socketId, stream } });
-  }
-};
-
-export const removeStream = (socketId) => (dispatch, getState) => { // TODO no need for a thunk
-  if (socketId) {
-    dispatch({ type: REMOVE_STREAM, payload: { socketId } });
-  }
-};
-
-export const clearAllStreams = () => async (dispatch, getState) => {
-  dispatch({ type: CLEAR_ALL_STREAMS });
-};
-
-export const startLocalStream = () => async (dispatch, getState) => { // TODO move to localUser.actions and rename streamsReducer to remoteStreams.reducer
+export const startLocalStream = () => async (dispatch, getState) => {
   const { mediaConstraints, micEnabled, camEnabled } = getState().localUser;
 
   if (!micEnabled && !camEnabled) {
@@ -70,5 +55,37 @@ export const toggleLocalVideo = () => (dispatch, getState) => {
       localStream.getVideoTracks()[0].enabled = !localStream.getVideoTracks()[0].enabled;
       dispatch(addStream(socket.id, localStream));
     }
+  }
+};
+
+export const setMediaStreamConstraints = (constraints) => async (dispatch, getState) => {
+  await dispatch({ type: SET_LOCAL_USER_MEDIA_CONSTRAINTS, payload: { constraints } });
+
+  // Re-Start local stream to update constraints
+  const localStreamWrapper = getState().localUser.mediaStream;
+  if (localStreamWrapper) {
+    dispatch(startLocalStream());
+  }
+};
+
+export const toggleMicrophoneEnabled = () => (dispatch, getState) => {
+  const newEnabled = !getState().localUser.micEnabled;
+  dispatch({ type: SET_LOCAL_USER_MICROPHONE_ENABLED, payload: { enabled: newEnabled } });
+
+  // Re-Start local stream to update constraints
+  const localStreamWrapper = getState().localUser.mediaStream;
+  if (localStreamWrapper || newEnabled) {
+    dispatch(startLocalStream());
+  }
+};
+
+export const toggleCameraEnabled = () => (dispatch, getState) => {
+  const newEnabled = !getState().localUser.camEnabled;
+  dispatch({ type: SET_LOCAL_USER_CAMERA_ENABLED, payload: { enabled: newEnabled } });
+
+  // Re-Start local stream to update constraints
+  const localStreamWrapper = getState().localUser.mediaStream;
+  if (localStreamWrapper || newEnabled) {
+    dispatch(startLocalStream());
   }
 };
