+++ frontend/src/store/actions/localUser.actions.js
@@ -1,8 +1,8 @@
-import {INIT_LOCAL_USER_SOCKET, RESET_LOCAL_USER, SET_LOCAL_USER_LOADING, SET_LOCAL_USER_MEDIA_CONSTRAINTS, SET_LOCAL_USER_METADATA, SET_LOCAL_USER_MICROPHONE_ENABLED, SET_LOCAL_USER_CAMERA_ENABLED} from '../actionTypes';
+import {INIT_LOCAL_USER_SOCKET, RESET_LOCAL_USER, SET_LOCAL_USER_LOADING, SET_LOCAL_USER_METADATA} from '../actionTypes';
 import io from 'socket.io-client';
 import {getPersistedData, persistData, serializeQueryString} from '../../utils';
 import {addConnection, initDataChannelListeners, removeConnection} from './connections.actions';
-import {startLocalStream} from './streams.actions';
+import {startLocalStream} from './localStream.actions';
 import {setCurrentRoomError} from './room.actions';
 import {BACKEND_URL} from '../../constants';
 
@@ -100,34 +100,3 @@ export const setMetadata = (metadata) => (dispatch, getState) => {
   dispatch({ type: SET_LOCAL_USER_METADATA, payload: { metadata } });
 };
 
-export const setMediaStreamConstraints = (constraints) => async (dispatch, getState) => {
-  await dispatch({ type: SET_LOCAL_USER_MEDIA_CONSTRAINTS, payload: { constraints } });
-
-  // Re-Start local stream to update constraints
-  const localStreamWrapper = getState().localUser.mediaStream;
-  if (localStreamWrapper) {
-    dispatch(startLocalStream());
-  }
-};
-
-export const toggleMicrophoneEnabled = () => (dispatch, getState) => {
-  const newEnabled = !getState().localUser.micEnabled;
-  dispatch({ type: SET_LOCAL_USER_MICROPHONE_ENABLED, payload: { enabled: newEnabled } });
-
-  // Re-Start local stream to update constraints
-  const localStreamWrapper = getState().localUser.mediaStream;
-  if (localStreamWrapper || newEnabled) {
-    dispatch(startLocalStream());
-  }
-};
-
-export const toggleCameraEnabled = () => (dispatch, getState) => {
-  const newEnabled = !getState().localUser.camEnabled;
-  dispatch({ type: SET_LOCAL_USER_CAMERA_ENABLED, payload: { enabled: newEnabled } });
-
-  // Re-Start local stream to update constraints
-  const localStreamWrapper = getState().localUser.mediaStream;
-  if (localStreamWrapper || newEnabled) {
-    dispatch(startLocalStream());
-  }
-};
