+++ frontend/src/store/reducers/index.js
@@ -3,7 +3,7 @@ import roomReducer from './room.reducer';
 import localUserReducer from './localUser.reducer';
 import peersReducer from './peers.reducer';
 import connectionsReducer from './connections.reducer';
-import streamsReducer from './streams.reducer';
+import remoteStreamsReducer from './remoteStreams.reducer';
 import pluginReducer from './plugin.reducer';
 import chatReducer from './chat.reducer';
 
@@ -12,7 +12,7 @@ const reducer = combineReducers({
   localUser: localUserReducer,
   peers: peersReducer,
   connections: connectionsReducer,
-  streams: streamsReducer,
+  remoteStreams: remoteStreamsReducer,
   plugin: pluginReducer,
   chat: chatReducer
 });
